+++ tarantool/src/decimal.rs
@@ -6,12 +6,20 @@ use std::mem::size_of;
 use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
 
-/// A Decimal number implemented using the builtin tarantool api. **Note** that
-/// this api is not available in all versions of tarantool.
+/// A Decimal number implemented using the builtin tarantool api.
+
+/// ## Availability
+/// This api is not available in all versions of tarantool.
 /// Use [`tarantool::ffi::has_decimal`] to check if it is supported in your
 /// case.
-/// If `has_decimal` return `false`, using any function from this module
+/// If `has_decimal` returns `false`, using any function from this module
 /// will result in a **panic**.
+///
+/// ## Safety
+/// This API is not thread-safe by default. It sacrifices thread-safety for performance.
+/// Therfore using [`Decimal`] (even just instantiating it) in concurrent threads is an undefined behavior.
+///
+/// Use `thread_safe_decimal` crate feature to enable thread-safety by sacrificing some performance.
 ///
 /// [`tarantool::ffi::has_decimal`]: crate::ffi::has_decimal
 #[derive(Debug, Copy, Clone)]
@@ -118,15 +126,28 @@ impl Decimal {
         }
 
         let ndig = (self.precision() - self.scale() + scale as i32).max(1);
-        let mut ctx: Context = unsafe { &*CONTEXT }.clone();
-        ctx.set_precision(ndig as _).unwrap();
-        ctx.set_max_exponent(ndig as _).unwrap();
-        ctx.set_min_exponent(if scale != 0 { -1 } else { 0 })
-            .unwrap();
-        ctx.set_rounding(mode);
+        let mut update_ctx = |ctx: &mut dec::Context<_>| {
+            ctx.set_precision(ndig as _).unwrap();
+            ctx.set_max_exponent(ndig as _).unwrap();
+            ctx.set_min_exponent(if scale != 0 { -1 } else { 0 })
+                .unwrap();
+            ctx.set_rounding(mode);
 
-        ctx.plus(&mut self.inner);
-        check_status(ctx.status()).ok()?;
+            ctx.plus(&mut self.inner);
+            check_status(ctx.status()).ok()
+        };
+        #[cfg(not(feature = "thread_safe_decimal"))]
+        {
+            let mut ctx = unsafe { &*CONTEXT }.clone();
+            update_ctx(&mut ctx.0)?;
+        }
+        #[cfg(feature = "thread_safe_decimal")]
+        {
+            CONTEXT.with(|ctx| {
+                let mut ctx: Context = ctx.borrow().clone();
+                update_ctx(&mut ctx.0)
+            })?;
+        }
         Self::try_from(self.inner).ok()
     }
 
@@ -272,49 +293,54 @@ impl TryFrom<DecimalImpl> for Decimal {
 /// Context
 ////////////////////////////////////////////////////////////////////////////////
 
-type Context = dec::Context<DecimalImpl>;
-static mut CONTEXT: Lazy<Context> = Lazy::new(|| {
-    let mut ctx = Context::default();
-    ctx.set_rounding(dec::Rounding::HalfUp);
-    ctx.set_precision(ffi::DECIMAL_MAX_DIGITS as _).unwrap();
-    ctx.set_clamp(false);
-    ctx.set_max_exponent((ffi::DECIMAL_MAX_DIGITS - 1) as _)
-        .unwrap();
-    ctx.set_min_exponent(-1).unwrap();
-    ctx
-});
+#[derive(Clone)]
+struct Context(dec::Context<DecimalImpl>);
+
+impl Default for Context {
+    fn default() -> Self {
+        let mut ctx = dec::Context::default();
+        ctx.set_rounding(dec::Rounding::HalfUp);
+        ctx.set_precision(ffi::DECIMAL_MAX_DIGITS as _).unwrap();
+        ctx.set_clamp(false);
+        ctx.set_max_exponent((ffi::DECIMAL_MAX_DIGITS - 1) as _)
+            .unwrap();
+        ctx.set_min_exponent(-1).unwrap();
+        Self(ctx)
+    }
+}
+
+#[cfg(not(feature = "thread_safe_decimal"))]
+static mut CONTEXT: Lazy<Context> = Lazy::new(Context::default);
 
 // This will make Decimals thread safe in exchange for some performance penalty.
-// Seeing as how tarantool's decimals aren't thread safe, for now we don't care
-// thread_local! {
-//     static CONTEXT: Lazy<std::cell::RefCell<Context>> = Lazy::new(|| {
-//         let mut ctx = Context::default();
-//         ctx.set_rounding(dec::Rounding::HalfUp);
-//         ctx.set_precision(ffi::DECIMAL_MAX_DIGITS as _).unwrap();
-//         ctx.set_clamp(false);
-//         ctx.set_max_exponent((ffi::DECIMAL_MAX_DIGITS - 1) as _).unwrap();
-//         ctx.set_min_exponent(-1).unwrap();
-//         std::cell::RefCell::new(ctx)
-//     });
-// }
+#[cfg(feature = "thread_safe_decimal")]
+thread_local! {
+    static CONTEXT: Lazy<std::cell::RefCell<Context>> = Lazy::new(std::cell::RefCell::default);
+}
 
 #[inline(always)]
 fn with_context<F, T>(f: F) -> Option<T>
 where
-    F: FnOnce(&mut Context) -> T,
+    F: FnOnce(&mut dec::Context<DecimalImpl>) -> T,
 {
-    let ctx = unsafe { &mut CONTEXT };
-    let res = f(ctx);
-    let status = ctx.status();
-    ctx.set_status(Default::default());
-    check_status(status).map(|()| res).ok()
-    // CONTEXT.with(|ctx| {
-    //     let ctx = &mut *ctx.borrow_mut();
-    //     let res = f(ctx);
-    //     let status = ctx.status();
-    //     ctx.set_status(Default::default());
-    //     check_status(status).map(|()| res).ok()
-    // })
+    let update_ctx = |ctx: &mut dec::Context<_>| {
+        let res = f(ctx);
+        let status = ctx.status();
+        ctx.set_status(Default::default());
+        check_status(status).map(|()| res).ok()
+    };
+    #[cfg(not(feature = "thread_safe_decimal"))]
+    {
+        let ctx = unsafe { &mut CONTEXT };
+        update_ctx(&mut ctx.0)
+    }
+    #[cfg(feature = "thread_safe_decimal")]
+    {
+        CONTEXT.with(|ctx| {
+            let ctx = &mut *ctx.borrow_mut();
+            update_ctx(&mut ctx.0)
+        })
+    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -500,7 +526,7 @@ macro_rules! impl_bin_op {
             #[inline(always)]
             #[track_caller]
             fn $ass_op(&mut self, rhs: T) {
-                *self = self.$m(rhs).expect("overlow")
+                *self = self.$m(rhs).expect("overflow")
             }
         }
     };
@@ -649,8 +675,30 @@ macro_rules! impl_from_int {
 }
 
 impl_from_int! {i8 i16 i32 u8 u16 u32 => DecimalImpl::from}
-impl_from_int! {i64 isize => |num| CONTEXT.from_i64(num as _)}
-impl_from_int! {u64 usize => |num| CONTEXT.from_u64(num as _)}
+impl_from_int! {
+    i64 isize => |num| {
+        #[cfg(not(feature = "thread_safe_decimal"))]
+        {
+            CONTEXT.0.from_i64(num as _)
+        }
+        #[cfg(feature = "thread_safe_decimal")]
+        {
+            CONTEXT.with(|ctx| ctx.borrow_mut().0.from_i64(num as _))
+        }
+    }
+}
+impl_from_int! {
+    u64 usize => |num| {
+        #[cfg(not(feature = "thread_safe_decimal"))]
+        {
+            CONTEXT.0.from_u64(num as _)
+        }
+        #[cfg(feature = "thread_safe_decimal")]
+        {
+            CONTEXT.with(|ctx| ctx.borrow_mut().0.from_u64(num as _))
+        }
+    }
+}
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
 pub enum DecimalFromfloatError<T> {
@@ -862,6 +910,41 @@ mod test {
     use std::convert::TryFrom;
     use std::sync::Mutex;
     static DECIMALS_ARENT_THREAD_SAFE: Lazy<Mutex<()>> = Lazy::new(Default::default);
+
+    #[cfg(feature = "thread_safe_decimal")]
+    #[test]
+    fn thread_safe_decimal() {
+        let mut handles = Vec::new();
+        for _ in 0..100 {
+            handles.push(std::thread::spawn(move || {
+                for _ in 0..10 {
+                    // Somehow this combination of successful and erroneous parse
+                    // would consistently cause errors in the not thread-safe implementation.
+                    let _: Decimal = "-81.1e-1".parse().unwrap();
+                    let _ = "foobar".parse::<Decimal>().unwrap_err();
+                }
+            }));
+        }
+        for handle in handles {
+            handle.join().unwrap();
+        }
+    }
+
+    #[test]
+    fn from_string() {
+        let _lock = DECIMALS_ARENT_THREAD_SAFE.lock().unwrap();
+        let d: Decimal = "-81.1e-1".parse().unwrap();
+        assert_eq!(d.to_string(), "-8.11");
+        assert_eq!(decimal!(-81.1e-1).to_string(), "-8.11");
+
+        assert_eq!("foobar".parse::<Decimal>().ok(), None::<Decimal>);
+        assert_eq!("".parse::<Decimal>().ok(), None::<Decimal>);
+
+        // tarantool decimals don't support infinity or NaN
+        assert_eq!("inf".parse::<Decimal>().ok(), None::<Decimal>);
+        assert_eq!("infinity".parse::<Decimal>().ok(), None::<Decimal>);
+        assert_eq!("NaN".parse::<Decimal>().ok(), None::<Decimal>);
+    }
 
     #[test]
     fn from_num() {